move article and author to features
