+++ src/features/article/article-details/ArticleDetails.scss
